🔧 refactor(api): update Spotify API routes and pages to use dynamic auth options
